# :seedling: refactor SortForCreate to use sort.Slice