--- a/util/resource/resource_test.go
+++ b/util/resource/resource_test.go
@@ -17,10 +17,23 @@ limitations under the License.
 package resource
 
 import (
+	"math/rand"
 	"testing"
 
 	. "github.com/onsi/gomega"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+)
+
+var (
+	kindsWithPriority = []string{
+		"Namespace", "CustomResourceDefinition", "StorageClass", "PersistentVolume",
+		"PersistentVolumeClaim", "Secret", "ConfigMap", "ServiceAccount", "LimitRange",
+		"Pod", "ReplicaSet", "Endpoint",
+	}
+	kindsOther = []string{
+		"Deployment", "Service", "DaemonSet", "StatefulSet", "Job", "CronJob", "Ingress",
+		"NetworkPolicy", "Role", "RoleBinding",
+	}
 )
 
 func TestSortForCreate(t *testing.T) {
@@ -35,9 +48,49 @@ func TestSortForCreate(t *testing.T) {
 	ep := unstructured.Unstructured{}
 	ep.SetKind("Endpoint")
 
-	resources := []unstructured.Unstructured{ep, cm, ns}
+	dp := unstructured.Unstructured{}
+	dp.SetKind("Deployment")
+
+	ds := unstructured.Unstructured{}
+	ds.SetKind("DaemonSet")
+
+	resources := []unstructured.Unstructured{ds, dp, ep, cm, ns}
 	sorted := SortForCreate(resources)
-	g.Expect(sorted).To(HaveLen(3))
+	g.Expect(sorted).To(HaveLen(5))
 	g.Expect(sorted[0].GetKind()).To(BeIdenticalTo("Namespace"))
 	g.Expect(sorted[1].GetKind()).To(BeIdenticalTo("ConfigMap"))
+	g.Expect(sorted[2].GetKind()).To(BeIdenticalTo("Endpoint"))
+	// we have no way to determine the order of the last two elements
+	g.Expect(sorted[3].GetKind()).To(BeElementOf([]string{"DaemonSet", "Deployment"}))
+	g.Expect(sorted[4].GetKind()).To(BeElementOf([]string{"DaemonSet", "Deployment"}))
+}
+
+func TestSortForCreateAllShuffle(t *testing.T) {
+	g := NewWithT(t)
+
+	resources := make([]unstructured.Unstructured, 0, len(kindsWithPriority)+len(kindsOther))
+	for _, kind := range append(kindsWithPriority, kindsOther...) {
+		resource := unstructured.Unstructured{}
+		resource.SetKind(kind)
+		resources = append(resources, resource)
+	}
+	for j := 0; j < 100; j++ {
+		// determinically shuffle resources
+		rnd := rand.New(rand.NewSource(int64(j))) //nolint:gosec
+		rnd.Shuffle(len(resources), func(i, j int) {
+			resources[i], resources[j] = resources[j], resources[i]
+		})
+
+		sorted := SortForCreate(resources)
+		g.Expect(sorted).To(HaveLen(len(kindsWithPriority) + len(kindsOther)))
+		// first check that the first len(kindsWithPriority) elements are from
+		// the list of kinds with priority (and in order)
+		for i, res := range sorted[:len(kindsWithPriority)] {
+			g.Expect(res.GetKind()).To(BeIdenticalTo(kindsWithPriority[i]))
+		}
+		// while the rest of resources can be any of the other kinds
+		for _, res := range sorted[len(kindsWithPriority):] {
+			g.Expect(kindsOther).To(ContainElement(res.GetKind()))
+		}
+	}
 }