**What this PR does / why we need it**:

I found [SortForCreate](https://github.com/kubernetes-sigs/cluster-api/blob/4abf44cd85c4590602e4c10543d53cd4ec914845/util/resource/resource.go#L50) to be very inefficient, it's basically O(n^2), it also needlessly allocates. In order to keep the contract (even though this exported function [is not widely used](https://github.com/search?type=code&q=SortForCreate+language%3AGo&l=Go&p=1)), I didn't change this function's behaviour (i.e. its purity), ~~but I did deprecate it~~. (It could've been even faster if we opted for an in-place sort, but I didn't want to change the function's behaviour.)

The function doesn't appear to be used in any tight loops or with a large number of elements, so it's a low prio PR, but it's something that could bite people applying a large number of small resources (see how it falls off a cliff with 10k elements).

**Benchmark comparison**
```
name             old time/op    new time/op    delta
Sort/sort-10       10.0µs ±31%     3.7µs ±55%  -63.11%  (p=0.000 n=10+10)
Sort/sort-100       423µs ±10%      58µs ±28%  -86.35%  (p=0.000 n=10+10)
Sort/sort-1000     45.0ms ±13%     0.5ms ±12%  -98.83%  (p=0.000 n=10+9)
Sort/sort-10000     5.90s ±22%     0.01s ±30%  -99.86%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
Sort/sort-10         248B ± 0%      136B ± 0%  -45.16%  (p=0.000 n=10+10)
Sort/sort-100      2.04kB ± 0%    0.95kB ± 0%  -53.33%  (p=0.000 n=10+10)
Sort/sort-1000     25.2kB ± 0%     8.2kB ± 0%  -67.28%  (p=0.000 n=10+10)
Sort/sort-10000     358kB ± 0%      82kB ± 0%  -77.08%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
Sort/sort-10         5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.000 n=10+10)
Sort/sort-100        8.00 ± 0%      3.00 ± 0%  -62.50%  (p=0.000 n=10+10)
Sort/sort-1000       12.0 ± 0%       3.0 ± 0%  -75.00%  (p=0.000 n=10+10)
Sort/sort-10000      20.0 ± 0%       3.0 ± 0%  -85.00%  (p=0.000 n=10+10)
```

**GH issue**
Fixes #9253

/area util