@kokes 2023-08-21
I redid it a bit - I decided not to deprecate `SortForOrder` after all, because the `sort.Slice` usage is quite clunky (due to the closure that's required) and should only be used in a performance-critical scenario (and anyone can use it, should they need it). The new (allocating) implementation is good enough, so I'll leave it that way.

We could also:

1. Unexport `PriorityLess` to keep the API surface smaller.
2. Create a function to satisfy [SortFunc](https://pkg.go.dev/slices#SortFunc), though that API so super new that it won't be leveraged until 1.21 is the oldest Go version supported, which is probably quite far away atm.

Neither is a priority tho.