@richardcase 2023-08-24
@killianmuldoon @enxebre - do we care that the original behaviour doesn't match the comments in the code? I see we use it in the CRS and clusterctl (for cert-manager) code but i don't have any context on its original inclusion. 

(If we do care i'd say its out of scope for this PR though)