@kokes 2023-08-24
I've pushed two commits, where I:

- unexported `PriorityLess` since we don't use it anywhere othere than this package
- removed the benchmark since we don't run benchmarks in this projects apparently
- fixed the plural of `Pods` and `Endpoints`
- reworked the existing test to include elements with no priority to check they are assigned least priority
- added a new, more comprehensive test, where I add all sorts of resources, repeatedly shuffle and check that sorting these always results in all the kinds with priorities frontloaded and the rest of resources with no particular order